+++ DAL/Entities/SellingOrderDetail.cs
@@ -10,7 +10,7 @@ namespace DAL.Entities
     {
         [Key]
         public int SellOrderDetailID { get; set; }
-        public int StockCount { get; set; }
+        public float StockCount { get; set; }
 
         [Column(TypeName = "decimal(10,2)")]
         public decimal SellingPrice { get; set; }
